requirements.txt
amqp==5.0.9
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.6
django-cors-headers==3.8.0
django-jazzmin==2.4.8
django-tinymce==3.4.0
djangorestframework==3.12.4
idna==3.3
kombu==5.2.3
packaging==21.3
passlib==1.7.4
Pillow==8.3.2
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
PyJWT==1.7.1
pyparsing==3.0.7
python-decouple==3.4
pytz==2021.3
redis==4.1.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.2
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
gunicorn==20.1.0

